THM Capstone
In this final challenge of the vulnerability module, I was introduced to Fuel CMS hosted at a provided IP. After identifying the CMS version (1.4), I researched known exploits and successfully applied CVE-2018-16763 a Remote Code Execution vulnerability.

One of the key takeaways for me was using netcat for the first time to establish a listener and gain shell access. Setting up the listener, executing the exploit, and retrieving the flag from the /home/ubuntu directory felt like bringing together all the pieces I’ve learned.

This hands-on experience helped reinforce how vulnerabilities can be chained and exploited in the real world. It was insightful to walk through the exploit code, understand its behaviour, and watch everything come together live in the terminal.

💡 Every room adds another layer to my learning theory into practice. I'm not just learning what attackers can do, but how they do it.